Big Refector
Added Ram Node, moved a lot of header
specific functions to the header
